nat46: fix ICMPv6 error message dropped locally
This patch is propagated from the kernel 4.4 commit
1b96bd0e9ee9182566b119741854c03bf4b94a99

While routing IPv6 packets from a customer-side translated device (CLAT)
to a provider-side translated device (PLAT), it is possible that the IPv6
destination is unknown. In such a scenario, the IPv6 stack must send back
an ICMP error. However, the source IPv6 address of this error message does
not have a MAP-T translation. According to RFC2473, the translation layer
should use the tunnel's own IPv4 address for the IPv6 ICMP packet's source
address.

Change-Id: I784473cddf9214843c466d10763cb66852139ef6
Signed-off-by: Pavithra R <pavir@codeaurora.org>
Signed-off-by: Ken Zhu <quic_guigenz@quicinc.com>