netifd: fix ip lost after dhcp renew
When two more IPv4 addresses of the same subnetwork are assigned
to one device, if the primary address got deleted, the rest IP
addresses could be deleted as well.

To make sure that old IPv4 address deleted before assigning the
new address, both of them  must be proceeded in one time.

Change-Id: Ib989955c1511a14d34a5a45df1ad7e8713815b42
Signed-off-by: Ken Zhu <guigenz@codeaurora.org>
Signed-off-by: Ken Zhu <quic_guigenz@quicinc.com>