[qsdk]Add telemetry agent in open profile
Telemetry Agent module is to maintain statistics and data required
for the provisioning related actions. For this the telemetry agent
will be interacting with WiFi driver to collect raw data and
convert it to RM user-space processable format. This module will
be in kernel space. Telemetry is collected by WIFI Host and WIFI
FW from the in-band processing of MSDUs and PPDUs. WIFI driver
maintains the statistics in native form, Telemetry Agent (TA)
collects and maintains the processed stats. TA provides the
processed telemetry info to RM app using Relay FS

Change-Id: I413baecd0023b7ea7478874cf9bfac452654cec0
Signed-off-by: Mahendran P <quic_mahep@quicinc.com>