netifd: Fix patch rebase issue
This patch fixes patch rebase issues. Base version of netifd package is changed,
due to this some of the patches are wrongly applied to different places
and causing issues for example sae auth type is wrongly placed,
seeing failure in 6g radio bring-up.

And removed some of the duplicate changes.

Change-Id: Id57e85a61d4699d4ce3527a128f876fca1fdf9f5
Signed-off-by: Bhagavathi Perumal S <quic_bperumal@quicinc.com>