[base-files] Add generic debug mechanism and debug script to "tc-full"
The goal of this mechanism is to provide an easy and decentralized way
for a package to register a set of debugging commands, in the case where
someone wants to gather all the logs in the system and store them for
later analysis.

The applications can register a set of debug commands by adding a
script in /lib/debug. The "sysdebug" command will recursively execute
all these scripts, log their output and format it accordingly.

This debug script will show the qdisc status for all the network devices
registered in the system

Signed-off-by: ratchaya <quic_ratchaya@quicinc.com>
Change-Id: Id42f022961273c76a1f978428e7f5c13f33d3936